+++ app/src/main/java/com/example/firabasecrud/Obra.kt
@@ -7,7 +7,7 @@ data class Obra(
     var id : String? = "",
     var nombre : String? = "",
     var autor : String? = "",
-    var estrellas : Int? = 0,
+    var estrellas : Float? = 0f,
     var fecha : String? = Util.obtenerFecha(),
     var rutaImagen : String? = "",
     var idImagen : String? = ""
